Combine sgp and default utils.h, move into source/ with base classes
