+++ source/pgg_mode/PGGWorldSetup.cc
@@ -5,7 +5,7 @@
 #include "PGGHost.h"
 #include "PGGSymbiont.h"
 
-#include "../default_mode/utils.h"
+#include "../utils.h"
 
 /**
  * Input: The number of PGG hosts.
